working and moved to cpymadtools.tune
